moved getting user info to UserInfo, fix for editing googleuser
